Wip: Add RRM REST and datastore infrastructure concerning publications.
