+++ assets/js/modules/reader-revenue-manager/datastore/publications.js
@@ -183,29 +183,6 @@ const fetchUpdatePublicationStore = createFetchStore( {
 	isAction: true,
 } );
 
-const fetchGetTermsOfServiceStore = createFetchStore( {
-	baseName: 'getTermsOfService',
-	controlCallback: ( { tosURL } ) =>
-		get(
-			'modules',
-			MODULE_SLUG_READER_REVENUE_MANAGER,
-			'terms-of-service',
-			{ tosURL },
-			{ useCache: false }
-		),
-	reducerCallback: createReducer( ( state, termsOfService, { tosURL } ) => {
-		state.termsOfService = state.termsOfService || {};
-		state.termsOfService[ tosURL ] = termsOfService;
-	} ),
-	argsToParams: ( { tosURL } = {} ) => ( { tosURL } ),
-	validateParams: ( { tosURL } = {} ) => {
-		invariant(
-			typeof tosURL === 'string' && tosURL.length > 0,
-			'tosURL is required and must be a string.'
-		);
-	},
-} );
-
 const fetchGetSyncPublicationOnboardingStateStore = createFetchStore( {
 	baseName: 'getSyncPublicationOnboardingState',
 	controlCallback: ( { publicationID, publicationOnboardingState } ) =>
@@ -272,7 +249,6 @@ const fetchGetSyncPublicationOnboardingStateStore = createFetchStore( {
 const baseInitialState = {
 	publications: undefined,
 	publicationsByID: {},
-	termsOfService: {},
 };
 
 const baseActions = {
@@ -545,23 +521,6 @@ const baseResolvers = {
 			} );
 		}
 	},
-
-	*getTermsOfService( { tosURL } = {} ) {
-		if ( ! tosURL ) {
-			return;
-		}
-
-		const registry = yield commonActions.getRegistry();
-		const termsOfService = registry
-			.select( MODULES_READER_REVENUE_MANAGER )
-			.getTermsOfService( { tosURL } );
-
-		if ( termsOfService === undefined ) {
-			yield fetchGetTermsOfServiceStore.actions.fetchGetTermsOfService( {
-				tosURL,
-			} );
-		}
-	},
 };
 
 const baseSelectors = {
@@ -593,24 +552,6 @@ const baseSelectors = {
 		}
 
 		return state.publicationsByID?.[ publicationID ];
-	},
-
-	/**
-	 * Gets the Terms of Service HTML.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param {Object} state         Data store's state.
-	 * @param {Object} params        Terms of Service parameters.
-	 * @param {string} params.tosURL Terms of Service URL.
-	 * @return {(string|undefined)} Terms of Service HTML; `undefined` if not loaded.
-	 */
-	getTermsOfService( state, { tosURL } = {} ) {
-		if ( ! tosURL ) {
-			return undefined;
-		}
-
-		return state.termsOfService?.[ tosURL ];
 	},
 
 	/**
@@ -684,7 +625,6 @@ const store = combineStores(
 	fetchCreatePublicationStore,
 	fetchGetPublicationStore,
 	fetchUpdatePublicationStore,
-	fetchGetTermsOfServiceStore,
 	fetchGetSyncPublicationOnboardingStateStore,
 	{
 		initialState: baseInitialState,
